+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Adding download image function by [kienluu612@gmail.com].
+
 ## [r11.6.6] - 2021-06-11
 ### Changed
 - Publish FilterProduct.